Fix Date math for this week/last week.
According to this:
http://stackoverflow.com/questions/6356164/simple-javascript-date-math-not-really

You must always use the same Date object for setDate and getDate.

This was causing the date math to be wrong, and thus the dashboard couldn't
find any of the data files.

I don't know why it worked up until now, but this change appears to fix it.