Ignore subsequent calls to v2/telemetry.js' Telemetry.init().
If you have a long-lived script, you will have to reload the page so that
Telemetry.init can be called again when another date's data is available.

Without this, calling Telemetry.init over and over again could cause callbacks
to be called when the internal version->date structures are in an inconsistent
state.