Make update orphaning dashboard aware of the new Airflow-specific data location
We now schedule the weekly Spark job for the update orphaning dashboard
via Airflow. Airflow uses a new URL for the output files. Airflow also
operates on a one-week delay, so we have to drop the most recent date
as an option from the dropdown menu in the dashboard.

Note: older data files will be copied manually to the new location.