Accommodate new weekday when update orphaning Spark job is run
The update orphaning Spark job is now scheduled to run right after the
longitudinal dataset creation job on Sunday, instead of Wednesday. The
dates need to be adjusted as follows:

Monday: -5
Tuesday: -6
Wednesday: -7
Thursday: -8
Friday: -9
Saturday: -10
Sunday: -11

Note that this is in addition to the 7-day adjustment a few lines below
because of the fact that Airflow schedules these jobs with a one week
delay.