Choose initial dates more wisely now that we show latest versions.
When we switch to showing only the latest versions, we need to abandon some
heuristics for choosing
a) The first date we show upon loading
b) How much data we show in the crash plot

a) When we switch to a new version, adoption takes a while to tick up to the
level of just a few days ago. This means the previous assumption of "days are
only useful if their kuh is approx the same as a week ago" is invalid across
version boundaries. As such, we need a special rule to find a good day if there
is no such date in the current version.

I just chose "end of data minus one" since it seems to work.

b) The audience is required to be very sophisticated to derive any sense out of
this dashboard at all, so hiding the end of the crash graph because it will
always be a spike is not necessary. Since it was using a similar heuristic to
a), above, the simplest way to get around this is through: by just showing
everything.