bug 1349536 - Add "Other" channel to crashdash
With users of old versions of Windows having been migrated to ESR in 52, it
suddenly becomes interesting to look at ESR crash rates and adoption.

Sadly ESR doesn't fit into the nice "latest versions" mechanism, so for now it
is reporting the most-used version. This works well for the latest ESR because
of the aforementioned influx of users, but may need to be tweaked in time for
the next one.