[evo] No longer hit the server several times to determine kind and buckets.
Previously, when changing the measure on evo.html, we would update the selected
aggregates first before pulling down the new evolution data. This would ensure
switching from, say, linear to flag, would mean we would stop displaying the
25th percentile and start displaying false, true, and invalid value percentages.

This was accomplished through calling Telemetry.getHistogramInfo for the
measure on every displayed channel and version, with no filters.

This hit the aggregates.tmo endpoint, asking for the full submission_date
aggregate data for the entire display... which was then ignored except for the
metric's `kind` and `buckets`.

To add insult to injury, this was unlikely to be cached, since we asked for it
filter-free. Usually there is at least an `application=Firefox` filter in place
which means we ask for all of this data, then never reuse it.

This change skips this entirely and pulls the kind and buckets out of the
result we get from asking for the display data.

This shaves anywhere from two to twenty-two seconds off of the time it takes to
switch metrics on evo.html.