Telemetry.getVersions should return [] if from and to have different channels.
Telemetry.getVersions will lexicographically filter candidate channel+version
pairs based on fromVersion and toVersion.

If fromVersion is the default min_channel_version (which when null is taken to
be the 3rd-most-recent nightly version) it works acceptably so long as the
max_channel_version starts with a letter in the alphabet before "n".

With release-channel aggregates coming back, "r" isnt' before "n" so this bug
becomes a problem.