Support auto-login when trying to load app state with release channels.
When being shared a link pointing at release data, when moving from evo to dist
or vice versa, or when simply reloading the page a user will no longer be
authorized to view release data but their state will have release-channel state
in it.

In this case we'll pop out to the authorization flow automatically as though
authorization had failed.

In most cases I expect this to immediately redirect back to the app in-state.